+++ NEWS.md
@@ -2,6 +2,7 @@
 
 * Modified the default vertical positioning of column and panel headings in forest_plot()
 * Added argument panel.headings.align to forest_plot() to control centring of panel headings.
+* Added argument col.heading.rule to forest_plot() to add a horizontal rule below column headings.
 
 # ckbplotr 0.11.2